Restructure package. No longer use a extension trait to match std.
Still outstanding: How do I hide the concrete structs from the trait?
