+++ src/stream/double_ended_stream/rfind.rs
