+++ src/stream/double_ended_stream/rfold.rs
