+++ src/stream/double_ended_stream/try_rfold.rs
